update security configurations to set user in session on successful login as part of auth flow
